Master

+++ src/Controller/UtilisateurController.php
@@ -3,6 +3,8 @@
 namespace App\Controller;
 
 use App\Entity\User;
+use App\Repository\StatusRepository;
+use App\Repository\TaskRepository;
 use App\Repository\UserRepository;
 use Doctrine\ORM\Mapping\Id;
 use Symfony\Component\HttpFoundation\Request;
@@ -44,11 +46,25 @@ class UtilisateurController extends AbstractController
     /**
     * @Route("/utilisateur/{id}/delete", name="app_user_delete", methods={"POST"})
     */
-    public function deleteUser(Request $request, User $user, UserRepository $userRepository): Response
+    public function deleteUser(Request $request, User $user, UserRepository $userRepository, StatusRepository $repoStatut, TaskRepository $repoTask): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->request->get('_token'))) {
+        
+        $statut_tache_user = $repoStatut->find(2);
+        $nb_tache_en_cours = $repoTask->CountTask($user,$statut_tache_user);
+
+        if ( count($nb_tache_en_cours) == 0 ){
+            // dd($nb_tache_en_cours);
+            
+           if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->request->get('_token'))) {
             $userRepository->remove($user);
+            }
+
+        }else{
+            $this->addFlash("failed", "Impossible, tâches en cours pour l'utilisateur ");
+            return $this->redirectToRoute('app_utilisateur');
         }
+        
+        
 
         return $this->redirectToRoute('app_utilisateur', [], Response::HTTP_SEE_OTHER);
     }
